extra tests

+++ support/cas-server-support-hazelcast-monitor/src/main/java/org/apereo/cas/monitor/HazelcastHealthIndicator.java
@@ -82,7 +82,7 @@ public class HazelcastHealthIndicator extends AbstractCacheHealthIndicator {
         }
 
         @Override
-        public int getPercentFree() {
+        public long getPercentFree() {
             final long capacity = getCapacity();
             if (capacity == 0) {
                 return 0;
